get auth service to work locally with SQLite DB - no google
