+++ .gitignore
@@ -36,6 +36,9 @@ bld/
 [Ll]og/
 [Ll]ogs/
 
+# SQLite Generated files
+SQLite/
+
 # Visual Studio 2015/2017 cache/options directory
 .vs/
 # Uncomment if you have tasks that create the project's static files in wwwroot
