+++ AuthenticationService/MicroManagement.Auth.WebAPI/MicroManagement.Auth.WebAPI.csproj
@@ -26,4 +26,6 @@
     <ProjectReference Include="..\MicroManagement.Auth.DataContracts\MicroManagement.Auth.DataContracts.csproj" />
   </ItemGroup>
 
+  <ProjectExtensions><VisualStudio><UserProperties properties_4launchsettings_1json__JsonSchema="" /></VisualStudio></ProjectExtensions>
+
 </Project>
