+++ AuthenticationService/MicroManagement.Auth.WebAPI/Startup/SqliteInitializationService.cs
@@ -1,0 +1,28 @@
+﻿
+using MicroManagement.Auth.WebAPI.Persistence;
+using Microsoft.EntityFrameworkCore;
+
+namespace MicroManagement.Auth.WebAPI
+{
+    /// <summary>
+    /// Initializes the SQLite DB in case it's not
+    /// </summary>
+    public class SqliteInitializationService(IDbContextFactory<AuthenticationServiceDbContext> _dbContextFactory) : IHostedService
+    {
+        /// <inheritdoc/>
+        public async Task StartAsync(CancellationToken cancellationToken)
+        {
+            using var dbContext = await _dbContextFactory.CreateDbContextAsync();
+
+            if (!dbContext.Database.IsSqlite())
+                throw new InvalidOperationException("WARNING, This service should only be injected for Local Development on SQLite DB!");
+
+            await dbContext.Database.EnsureCreatedAsync();
+        }
+
+        /// <inheritdoc/>
+        public Task StopAsync(CancellationToken cancellationToken)
+            => Task.CompletedTask;
+
+    }
+}
