+++ AuthenticationService/MicroManagement.Auth.WebAPI/appsettings.Development.json
@@ -4,5 +4,9 @@
       "Default": "Trace",
       "Microsoft.AspNetCore": "Trace"
     }
+  },
+  "Jwt": {
+    "JwtAccessKey": "SomeFakeAccessKeyDoNotTryToStealItYouWontDoAnythingWithItNotUsedInProdDummy",
+    "JwtRefreshKey": "SameHere.NoWayYouCanUseThisToStealMyData.MaybeIfYouStoleMyPCLol",
   }
 }
